Move custom code in LyoD services under .xtra
* PEM Svc
* Location Svc
* Robot Twin
* Shelf Twin
