--- lyo-services/webapp-twin-shelf/src/main/java/se/ericsson/cf/scott/sandbox/twins/shelf/manager/LeshanManager.java
@@ -1,20 +1,0 @@
-package se.ericsson.cf.scott.sandbox.twins.shelf.manager;
-
-/**
- * TODO
- *
- * @version $version-stub$
- * @since   TODO
- */
-public class LeshanManager {
-    public static void buildTheLwm2mServer() {
-//        final LeshanServer server = new LeshanServerBuilder().build();
-//
-//        // listen for new client registrations
-//        server.getClientRegistry().addListener(new ScottClientRegistryListener(server));
-//        // listen for observe notifications
-//        server.getObservationRegistry().addListener(new ScottObservationRegistryListener());
-//
-//        server.start();
-    }
-}
