--- lyo-services/webapp-twin/src/main/java/se/ericsson/cf/scott/sandbox/twin/ServiceProviderRepository.java
@@ -1,17 +1,0 @@
-package se.ericsson.cf.scott.sandbox.twin;
-
-import java.util.Collection;
-import org.eclipse.lyo.oslc4j.core.model.ServiceProvider;
-
-/**
- * TODO
- *
- * @since   TODO
- */
-public interface ServiceProviderRepository {
-    Collection<ServiceProvider> getServiceProviders();
-
-    ServiceProvider[] asSpArray();
-
-    void addServiceProvider(ServiceProvider sp);
-}
