--- lyo-services/webapp-twin/src/main/java/se/ericsson/cf/scott/sandbox/twin/trs/LyoStoreManager.java
@@ -1,21 +1,0 @@
-package se.ericsson.cf.scott.sandbox.twin.trs;
-
-import java.io.IOException;
-import org.apache.jena.sparql.ARQException;
-import org.eclipse.lyo.store.Store;
-import org.eclipse.lyo.store.StoreFactory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import static se.ericsson.cf.scott.sandbox.twin.TwinAdaptorHelper.p;
-
-/**
- * TODO
- *
- * @version $version-stub$
- * @since   TODO
- */
-public class LyoStoreManager {
-    private final static Logger log = LoggerFactory.getLogger(LyoStoreManager.class);
-
-}
