--- lyo-services/webapp-twin/src/main/java/se/ericsson/cf/scott/sandbox/twin/trs/MqttManager.java
@@ -1,15 +1,0 @@
-package se.ericsson.cf.scott.sandbox.twin.trs;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-/**
- * TODO
- *
- * @version $version-stub$
- * @since   TODO
- */
-public class MqttManager {
-    private final static Logger log = LoggerFactory.getLogger(MqttManager.class);
-
-}
