+++ lyo-services/webapp-twin/src/main/java/se/ericsson/cf/scott/sandbox/twin/xtra/trs/LyoStoreManager.java
@@ -1,0 +1,15 @@
+package se.ericsson.cf.scott.sandbox.twin.xtra.trs;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * TODO
+ *
+ * @version $version-stub$
+ * @since   TODO
+ */
+public class LyoStoreManager {
+    private final static Logger log = LoggerFactory.getLogger(LyoStoreManager.class);
+
+}
